# Reduce memory allocations in bucketBlock.readChunkRange()